pyproject.toml
[project]
name = "PaNETExtractor"
version = "0.1.0"
requires-python = ">=3.10"
description = "Extract paNET terms from text"
readme = "README.md" 
dependencies = [
 "fastapi[standard]>=0.115.13",
 "httpx>=0.28.1",
 "panet_technique_matcher",
 "data_provider",
 "techniques_extractor",
 "python-dotenv>=1.1.1",
 "transformers[torch]>=4.52.4",
 "logging>=0.4.9.6",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.12.5",
]


[tool.uv.sources]
panet_technique_matcher = {workspace = true}
techniques_extractor = {workspace = true}
data_provider = {workspace = true}

[tool.uv.workspace]
members = [
    "packages/*",
    "premier",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[[tool.uv.index]]
name = "testpypi" 
url  = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true
